完成file的增删改查和测试

+++ db/migration/000001_init_schema.up.sql
@@ -1,10 +1,10 @@
 CREATE TABLE `files` (
-  `id` integer PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint PRIMARY KEY AUTO_INCREMENT,
   `file_sha1` varchar(255) UNIQUE NOT NULL COMMENT '文件的sha1值',
   `file_name` varchar(255) NOT NULL COMMENT '文件名',
   `file_size` bigint NOT NULL DEFAULT 0 COMMENT '文件大小',
   `file_addr` varchar(1024) NOT NULL DEFAULT '' COMMENT '文件存储地址',
-  `status` int NOT NULL DEFAULT 0 COMMENT '文件状态(0: 正常,1:禁用, 2: 删除)',
+  `status` tinyint NOT NULL DEFAULT 0 COMMENT '文件状态(0: 正常,1:禁用, 2: 删除)',
   `create_at` timestamp NOT NULL DEFAULT (CURRENT_TIMESTAMP) COMMENT '创建时间',
   `update_at` timestamp NOT NULL DEFAULT (CURRENT_TIMESTAMP) COMMENT '更新时间',
   `extend` int NOT NULL DEFAULT 0 COMMENT '扩展字段',
